+++ lib/ember/appkit.rb
@@ -1,5 +1,0 @@
-require 'ember-rails'
-require 'ember/source'
-require 'es6_module_transpiler/rails'
-
-module Ember::Appkit; end