+++ lib/generators/ember/scaffold_override.rb
@@ -1,0 +1,2 @@
+require 'rails/generators'
+require 'rails/generators/rails/scaffold/scaffold_generator'